requirements.txt
Flask==1.0.2
gspread==3.1.0
oauth2client==4.1.3
pytz
google-cloud-datastore==1.8.0
pytest
